can now save friends by checking off the user name. The friends are beind saved on the backend
