+++ GolfProfile/Base.lproj/Main.storyboard
@@ -589,10 +589,10 @@
             </objects>
             <point key="canvasLocation" x="1863" y="-1142"/>
         </scene>
-        <!--Add Friends View Controller-->
+        <!--Edit Friends View Controller-->
         <scene sceneID="Z6a-zD-OQk">
             <objects>
-                <viewController id="2nK-tA-FpD" customClass="AddFriendsViewController" customModule="GolfProfile" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="2nK-tA-FpD" customClass="EditFriendsViewController" customModule="GolfProfile" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="3NS-xm-Q5i"/>
                         <viewControllerLayoutGuide type="bottom" id="d8P-tw-a6j"/>
@@ -601,19 +601,19 @@
                         <rect key="frame" x="0.0" y="0.0" width="320" height="568"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <tableView clipsSubviews="YES" contentMode="scaleToFill" fixedFrame="YES" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="122" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="5Uq-Ns-glz">
-                                <rect key="frame" x="0.0" y="122" width="320" height="446"/>
+                            <tableView clipsSubviews="YES" contentMode="scaleToFill" fixedFrame="YES" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="82" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="5Uq-Ns-glz">
+                                <rect key="frame" x="0.0" y="79" width="320" height="489"/>
                                 <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
                                 <prototypes>
-                                    <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" selectionStyle="none" indentationWidth="10" reuseIdentifier="addFriendCell" rowHeight="122" id="8v2-v2-Zm1" customClass="AddFriendCell" customModule="GolfProfile" customModuleProvider="target">
-                                        <rect key="frame" x="0.0" y="92" width="320" height="122"/>
+                                    <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" selectionStyle="none" indentationWidth="10" reuseIdentifier="editFriendCell" rowHeight="82" id="8v2-v2-Zm1" customClass="EditFriendCell" customModule="GolfProfile" customModuleProvider="target">
+                                        <rect key="frame" x="0.0" y="92" width="320" height="82"/>
                                         <autoresizingMask key="autoresizingMask"/>
                                         <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="8v2-v2-Zm1" id="u5y-ri-7lA">
-                                            <rect key="frame" x="0.0" y="0.0" width="320" height="121"/>
+                                            <rect key="frame" x="0.0" y="0.0" width="320" height="81"/>
                                             <autoresizingMask key="autoresizingMask"/>
                                             <subviews>
                                                 <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" fixedFrame="YES" text="UserName" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="TXw-ye-pxj">
-                                                    <rect key="frame" x="15" y="50" width="274" height="21"/>
+                                                    <rect key="frame" x="74" y="8" width="175" height="21"/>
                                                     <color key="backgroundColor" red="1" green="1" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
                                                     <fontDescription key="fontDescription" type="system" pointSize="17"/>
                                                     <color key="textColor" cocoaTouchSystemColor="darkTextColor"/>
@@ -632,13 +632,6 @@
                                     <outlet property="delegate" destination="2nK-tA-FpD" id="ROT-Qt-XE0"/>
                                 </connections>
                             </tableView>
-                            <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="aVZ-P7-gfD">
-                                <rect key="frame" x="131" y="72" width="58" height="30"/>
-                                <state key="normal" title="RELOAD"/>
-                                <connections>
-                                    <action selector="reload:" destination="2nK-tA-FpD" eventType="touchUpInside" id="0Vf-pU-a6C"/>
-                                </connections>
-                            </button>
                         </subviews>
                         <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
                     </view>
@@ -654,7 +647,6 @@
         <!--Navigation Controller-->
         <scene sceneID="mud-z2-2vx">
             <objects>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="WGs-FC-Pql" userLabel="First Responder" sceneMemberID="firstResponder"/>
                 <navigationController automaticallyAdjustsScrollViewInsets="NO" id="2Ob-IZ-z3K" sceneMemberID="viewController">
                     <toolbarItems/>
                     <navigationBar key="navigationBar" contentMode="scaleToFill" id="gMQ-Al-yQN">
@@ -666,6 +658,7 @@
                         <segue destination="fnr-Zv-jKV" kind="relationship" relationship="rootViewController" id="Tqg-Js-hDD"/>
                     </connections>
                 </navigationController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="WGs-FC-Pql" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="438" y="-2413"/>
         </scene>
